requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
cachetools==7.0.5
cassandra-driver==3.29.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.102.0
docker==7.1.0
entrypoints==0.4
fastapi==0.135.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
geomet==1.1.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
gunicorn==21.2.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kafka-python==2.3.0
kiwisolver==1.5.0
Mako==1.3.10
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
numpy==1.26.4
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==23.2
pandas==2.3.3
pillow==12.1.1
prettytable==3.17.0
prometheus-fastapi-instrumentator==6.1.0
prometheus_client==0.24.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==15.0.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2023.4
PyYAML==6.0.3
querystring-parser==1.2.4
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
Werkzeug==3.1.6
zipp==3.23.0
